backend/requirements.txt
fastapi
python-dotenv
python-multipart
pydantic
uvicorn[standard]
langchain
langchain_google_genai